+++ src/main/java/everTale/everTale_be/auth/jwt/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package everTale.everTale_be.auth.jwt;
 
-import everTale.everTale_be.domain.user.domain.User;
+import everTale.everTale_be.domain.user.entity.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import lombok.Getter;
