+++ src/main/java/everTale/everTale_be/auth/service/AuthService.java
@@ -7,8 +7,8 @@ import everTale.everTale_be.auth.dto.response.UserInfoResponseDto;
 import everTale.everTale_be.auth.jwt.JwtProvider;
 import everTale.everTale_be.auth.jwt.JwtUtil;
 import everTale.everTale_be.domain.profile.util.UserHelper;
-import everTale.everTale_be.domain.user.domain.Enum.LoginProvider;
-import everTale.everTale_be.domain.user.domain.User;
+import everTale.everTale_be.domain.user.entity.Enum.LoginProvider;
+import everTale.everTale_be.domain.user.entity.User;
 import everTale.everTale_be.domain.user.repository.UserRepository;
 import everTale.everTale_be.global.apiPayload.code.status.ErrorStatus;
 import everTale.everTale_be.global.apiPayload.exception.handler.BadRequestHandler;
