--- src/main/java/everTale/everTale_be/domain/profile/domain/Enum/ProfileType.java
@@ -1,5 +1,0 @@
-package everTale.everTale_be.domain.profile.domain.Enum;
-
-public enum ProfileType {
-    PARENT, CHILD
-}
