+++ src/main/java/everTale/everTale_be/domain/profile/dto/ProfileSummary.java
@@ -1,6 +1,6 @@
 package everTale.everTale_be.domain.profile.dto;
 
-import everTale.everTale_be.domain.profile.domain.Profile;
+import everTale.everTale_be.domain.profile.entity.Profile;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
