+++ src/main/java/everTale/everTale_be/domain/profile/dto/response/ProfileEnterResponseDto.java
@@ -1,7 +1,7 @@
 package everTale.everTale_be.domain.profile.dto.response;
 
-import everTale.everTale_be.domain.profile.domain.Enum.ProfileType;
-import everTale.everTale_be.domain.profile.domain.Profile;
+import everTale.everTale_be.domain.profile.entity.Enum.ProfileType;
+import everTale.everTale_be.domain.profile.entity.Profile;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
