+++ src/main/java/everTale/everTale_be/domain/profile/entity/Enum/ProfileType.java
@@ -1,0 +1,5 @@
+package everTale.everTale_be.domain.profile.entity.Enum;
+
+public enum ProfileType {
+    PARENT, CHILD
+}
