+++ src/main/java/everTale/everTale_be/domain/profile/service/ProfileService.java
@@ -3,8 +3,8 @@ package everTale.everTale_be.domain.profile.service;
 import everTale.everTale_be.auth.jwt.JwtProvider;
 import everTale.everTale_be.auth.jwt.JwtUtil;
 import everTale.everTale_be.auth.service.TokenAuthService;
-import everTale.everTale_be.domain.profile.domain.Enum.ProfileType;
-import everTale.everTale_be.domain.profile.domain.Profile;
+import everTale.everTale_be.domain.profile.entity.Enum.ProfileType;
+import everTale.everTale_be.domain.profile.entity.Profile;
 import everTale.everTale_be.domain.profile.dto.request.ChildProfileRequestDto;
 import everTale.everTale_be.domain.profile.dto.request.ChildProfileUpdateRequestDto;
 import everTale.everTale_be.domain.profile.dto.request.ParentProfileUpdateRequestDto;
@@ -12,7 +12,7 @@ import everTale.everTale_be.domain.profile.dto.request.PasswordUpdateRequestDto;
 import everTale.everTale_be.domain.profile.dto.response.*;
 import everTale.everTale_be.domain.profile.repository.ProfileRepository;
 import everTale.everTale_be.domain.profile.util.UserHelper;
-import everTale.everTale_be.domain.user.domain.User;
+import everTale.everTale_be.domain.user.entity.User;
 import everTale.everTale_be.domain.user.repository.UserRepository;
 import everTale.everTale_be.global.apiPayload.code.status.ErrorStatus;
 import everTale.everTale_be.global.apiPayload.exception.handler.BadRequestHandler;
