+++ src/main/java/everTale/everTale_be/domain/user/entity/Enum/LoginProvider.java
@@ -1,4 +1,4 @@
-package everTale.everTale_be.domain.user.domain.Enum;
+package everTale.everTale_be.domain.user.entity.Enum;
 
 public enum LoginProvider {
     LOCAL, NAVER, KAKAO,
